chore: los restos que quedaban sueltos
sum() acumulaba en float. De el cuelgan mean() y mse_loss, o sea el
gradiente inicial de casi cualquier backward, y clip_grad_norm ya usaba
double justo al lado por este mismo motivo. Ahora tambien.

device_free() se tragaba el codigo de retorno. No puede lanzar —lo llama
~Storage()— pero si avisar: hoy una doble liberacion es invisible y el
error reaparece pegado a la siguiente llamada de CUDA, que no tiene la
culpa. Un fprintf la primera vez y a limpiar, igual que con los kernels.

El job de CUDA compilaba y ahi se quedaba. La paridad se omite sola
cuando no hay dispositivo (test_cuda_parity.cpp:139), asi que ejecutar
ctest en el runner sin GPU si aporta: prueba la biblioteca entera
compilada con ENGINE_CUDA=ON, que es el camino de respaldo, mas
test_cuda_indexing, que replica los indices de los kernels en CPU.

Y el README.es.md, que decia 368 comprobaciones (524, o 589 con tarjeta),
llamaba "planificado" al backend CUDA doce lineas antes de marcar la
fase 6 hecha, y listaba tres trabajos de CI de los seis que hay.

Verificado: 524 sin CUDA y 589 con, ambos en verde.

+++ src/tensor.cpp
@@ -976,10 +976,14 @@ Tensor Tensor::softmax() const {
 
 // Reducción Suma a un escalar {1}
 Tensor Tensor::sum() const {
-    float total = 0.0f;
+    // El acumulador es double aunque los datos sean float, igual que en
+    // clip_grad_norm. Sumar un millon de valores en float pierde los bits bajos
+    // en cuanto el total crece frente a cada sumando, y de esta reduccion cuelga
+    // mean(), o sea mse_loss y el gradiente inicial de cualquier backward.
+    double total = 0.0;
     for (float v : data()) total += v;
     bool req_g = track(requires_grad());
-    Tensor res({1}, std::vector<float>{total}, req_g);
+    Tensor res({1}, std::vector<float>{static_cast<float>(total)}, req_g);
 
     if (req_g) {
         res.impl_->parents = { impl_ };
